+++ src/lottie/card.ts
@@ -1,0 +1,17 @@
+import { Entity } from "@galacean/engine";
+import { LottieAnimation } from "@galacean/engine-lottie";
+
+
+export default function lottieFunc(rootEntity: Entity, lottieEntity: Entity): Entity{
+
+    lottieEntity.transform.setPosition(0, 0, 16);
+    rootEntity.addChild(lottieEntity);
+    const lottie = lottieEntity.getComponent(LottieAnimation);
+
+    if(lottie instanceof LottieAnimation){
+        lottieEntity.transform.setScale(0.2, 0.2, 0.2);
+        lottie.play();
+    }
+
+    return lottieEntity;
+}