+++ src/utils/url.ts
@@ -1,7 +1,12 @@
 // get static model resources
-function getModelFileUrl(path: string, modelName: string, fileType: string): string {
-    return `${path}models/${modelName}/${modelName}.${fileType}`;
+function getFileUrl(type: string, path: string, modelName: string, fileType: string): string {
+    return `${path}/${type}/${modelName}/${modelName}.${fileType}`;
 }
 
 
-export default getModelFileUrl
+function getSceneFileUrl(path: string, modelName: string, fileName: string, fileType: string): string {
+    return `${path}/scene/${modelName}/${fileName}.${fileType}`;
+}
+
+
+export { getFileUrl, getSceneFileUrl }